Rename encryptMessage files
