+++ src/commands/encryptMessage.js
