+++ test/commands/encryptMessage.js
@@ -13,7 +13,7 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-import encrypt from '../../src/commands/encrypt';
+import encrypt from '../../src/commands/encryptMessage';
 import cryptoModule from '../../src/utils/cryptoModule';
 import * as input from '../../src/utils/input';
 import * as print from '../../src/utils/print';
